+++ src/Util/util.h
@@ -20,7 +20,6 @@
 #include <vector>
 #include <atomic>
 #include <unordered_map>
-#include "TimeTicker.h"
 #if defined(_WIN32)
 #undef FD_SETSIZE
 //修改默认64为1024路
@@ -172,52 +171,6 @@ private:
         static std::atomic<size_t> instance(0); \
         return instance; \
     }
-
-class BytesSpeed {
-public:
-    BytesSpeed() = default;
-    ~BytesSpeed() = default;
-
-    /**
-     * 添加统计字节
-     */
-    BytesSpeed &operator+=(size_t bytes) {
-        _bytes += bytes;
-        if (_bytes > 1024 * 1024) {
-            //数据大于1MB就计算一次网速
-            computeSpeed();
-        }
-        return *this;
-    }
-
-    /**
-     * 获取速度，单位bytes/s
-     */
-    int getSpeed() {
-        if (_ticker.elapsedTime() < 1000) {
-            //获取频率小于1秒，那么返回上次计算结果
-            return _speed;
-        }
-        return computeSpeed();
-    }
-
-private:
-    int computeSpeed() {
-        auto elapsed = _ticker.elapsedTime();
-        if (!elapsed) {
-            return _speed;
-        }
-        _speed = (int)(_bytes * 1000 / elapsed);
-        _ticker.resetTime();
-        _bytes = 0;
-        return _speed;
-    }
-
-private:
-    int _speed = 0;
-    size_t _bytes = 0;
-    Ticker _ticker;
-};
 
 std::string makeRandStr(int sz, bool printable = true);
 std::string hexdump(const void *buf, size_t len);
